requirements.txt
streamlit
Flask
flask-cors
requests
python-dotenv
PyPDF2
google-generativeai
gunicorn
